pythonPackages.uvicorn: fix dependency on old uvloop
Before this change, uvicorn required an older version of uvloop than was
present. The latest published version (0.9.0) actually picks up the version we
have, but that also bumps websockets to a later version, and that looks like
an API-breaking change. So the least harm is just to smooth out the uvloop
hardcoded version.

This has been tested (as in, getting a basic web server up) with the three
pythonPackages that depend on this (fastapi, sanic, and starlette).